chore: move Book type to separate file
