+++ src/libs/shared/src/types/best-sellers.ts
@@ -1,3 +1,5 @@
+import { Book } from "./book";
+
 export type BestSellers = {
   bestSellers: BestSeller[];
 };
@@ -8,21 +10,4 @@ export type BestSeller = {
   displayName: string;
   image: string;
   books: Book[];
-};
-
-export type Book = {
-  title: string;
-  author: string;
-  image: string;
-  description: string;
-  isbn13: string;
-  isbn10: string;
-  price: number;
-  priceUnit: string;
-  publisher: string;
-  categories?: string[];
-  createdAt: Date;
-  updatedAt: Date;
-  rank?: number;
-  ageGroup?: string;
 };
